fix: use ARGB int constructor for ComposeColor to avoid invalid colorspace index

+++ zero-ui/src/main/java/com/hluhovskyi/zero/common/Color.kt
@@ -6,5 +6,5 @@ fun ColorValue.toCompose(): ComposeColor =
     if (isUnspecified()) {
         ComposeColor.Unspecified
     } else {
-        ComposeColor(hex.toLong())
+        ComposeColor(hex.toInt())
     }
